Add hugo engine
Hugo engine build added via command executor. This means that hugo should be installed in the deploy environment. After a bit of trial I reached the conclusion that for the time being this should be the better solution. Hugo does not necessarly has a clear API so the command execution is much simpler and wraps some of the problems I was having using the API.

Issue #74